misc: rename config dir to lib
