+++ src/lib/nanoid.js
